Add proper error handling to the entire code
The reason why the error handling was ... poor previously was, that the
only way to get full stacktraces that I knew of was to panic after
getting the error. But this also means that it becomes impossible to
ignore errors, and you have to make sure that the lower level functions
where panic() is called are always called correctly if you don't want
the process to die.

However there is a library that allows annotating errors with
stacktraces. We wrap all "low level errors" (i.e. errors that come from
outside of iptsd) using that library, and pass them around as usual.
Once an error reaches the toplevel main() function it is logged and the
program is aborted.

This commit also includes a bit of cleanup in the various API wrappers
(uinput, ipts, epoll) to make them a bit less clunky.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>