Implement basic heatmap to singletouch processing
This is really basic at the moment, but what it does is enough to
successfully simulate a singletouch input.

This is an (attempted) implementation of qzed's prototype [1], but due
to some differences between his simulated data and the actual heatmap,
and me being bad a maths, I changed some stuff. Since this is just a
prototype, it can be refactored later anyways.

Note 1: This currently has quirks for SB2 hardcoded (Y axis being mirrored).
Note 2: The input is *very* jittery. We can add smoothing to it, but I
	don't assume anyone wants to draw with this, so it might not be a big
	issue

[1]: https://github.com/qzed/touch-detection-prototype

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>