Multitouch!
So, the previous commit already had everything required to implement
(basic) multitouch support - the calculations were all there and worked.
But I was too lazy to implement a multitouch device so I reused the
singletouch device that already existed and limited the input to one
contact point.

Begone lazyness!

This is still not full 4.19 level multitouch support, but I guess it
should be pretty useable. Whats still missing is mostly support for
contact area / pressure, and noise filtering / detection.

And obviously this will still need adjustments for other IPTS devices,
it has only been tested on an SB2.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>