Finger tracking + some rework
The tracking still likes to freak out sometimes, since we don't have
noise filtering, and pressing on the screen hard enough can sometimes
count as 2 or 3 inputs. Still, it works way better than just using the
order of the fingers on the screen.

I want to apologize to anyone who reads this and has an idea what they
are doing, because I don't :P

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>