Reorganize the (post-)processing code
All the processing is now done in the processing module, without having
processing or global caches (except for the heatmap buffer) in main.

The basic processing for heatmaps was split up too and moved to
processing/heatmap. The code in processing itself is responsible for
finger tracking (which is not specific to the heatmap), and adjusting
the data for the touchscreen device.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>