Add config files for the different devices
These config files will contain device specific configuration such as
whether an axis should be inverted, or the size of the screen. The
screensize is used to properly calculate the resolution values of
the uinput devices.

Config files will be read from the following locations:
 - /usr/share/ipts
 - /usr/local/share/ipts
 - $PWD/config (for development)

Additionally /etc/ipts.conf will be parsed, as a way to override the
config that was automatically selected for the device.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>