Include the slot ID in finger tracking
When the index of a finger in a slot changes suddenly, libinput cannot
handle the event properly and throws an error. The finger input will
jump across the screen which can lead to touch gestures randomly
triggering.

The solution for this is to reuse the finger index as the slot ID, and
then fill up the remaining touch points with the indices that remain
free. The finger tracking already has most of the code to do that, it
just needs to be extended for the entire input array.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>