Update to the new UAPI interface
With the new interface, doesn't receive data through a single misc
device managed by the kernel anymore. Instead, it gets direct access to
every data buffer, as well as the doorbell. This means that instead of
polling the kernel who then polls the doorbell we can poll the doorbell
directly and simplify the code a bit.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>