Make the data parsing more safe
We currently don't do any checks on the incoming data, so when it is
corrupted iptsd will potentially read data that is way outside of the
bounds of the data buffer and crash with a segmentation fault.

By wrapping all pointer magic we can catch these out-of-bounds reads and
abort parsing.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>