Various tweaks to the meson build systemd
Meson is now able to automatically install the systemd files and the
example config. This can be disabled by setting -Dsystemd=false and
-Dsample_config=false. Building the debug tool can also be disabled
with -Ddebug_tool=false.

The various hardcoded paths in the config loader and the systemd file
are gone too now, they get populated automatically with the correct
values from the meson environment.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>