Overhaul stylus data parsing (again)
Thanks to @qzed's work on debugging the raw ipts data we figured out
that on gen5+ devices the stylus will send two paths to us, one through
the reports with, and one through those without a serial number. Mixing
these two paths will result in the jagged lines that have plagued ipts
for some time now.

Going forward the reports without a serial number will be discarded. The
data from reports that have a serial will be fully emitted (as opposed
to just sending the first element and skipping the rest).

To simplify the code, the reports from gen4 will be called v1 reports
from now on, while the reports from gen5+ will be called v2 reports. The
report type without a serial has been dropped from the protocol
definition to keep it simple and understandable.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>