Allow meson to configure and install openrc init script
The openrc init script can be installed by setting the meson option
-Dopenrc=true.

Additionally, the scripts for both systemd and openrc are now always
configured when you run meson (regardless of whether or not you use the
options -Dsystemd=true or -Dopenrc=true to install them). This change
allows the daemon to be more easily packaged for Gentoo.