daemon: Change how we store the active stylus
Instead of storing an index or a pointer to the active stylus, lets just define
that the last element in the styli vector is the currently active one. When we
switch the active stylus, we simply swap the new entry with the last one. And
when we add a new stylus it becomes the active one automatically.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>