container: Fix performance problem with _GLIBCXX_ASSERTIONS
Defining _GLIBCXX_ASSERTIONS will enable access checks in STL containers and in
unique_ptr. This is particulary noticable in the Image container, given how
often it is accessed during touch processing.

To improve performance with _GLIBCXX_ASSERTIONS we work around the access check
by using the raw pointer. However this will cause issues if the pointer is NULL,
because we just disabled the check for that.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>