ipts: Refactor heatmap dimension parsing
Apparently, the size of the heatmap is sent in the report before the one containing the data.
This means that the data at the end of the report is actually for the next report.

As a result, we can make the parsing a lot easier, since the dimension data is also needed for DFT processing