Revert "daemon: Read screen dimensions from DRM node instead of the config"
This reverts commit 068f5078b2411f08694841595c8f8063c9a7c65a.

As noted by @quo, newer devices support a HID report that will return the
dimensions of the screen, as well as information about invert_x / invert_y.

Given that we need the configs anyways for older devices, and the configs
allow for more precise values than DRM, it doesnt make sense to have three
different sources for screen width / height.