Add ContactFinder::reset()
This avoids throwing the contact finder and the blob detector away and
reallocating a mostly identical object (and in case of the advanced one
there is quite a lot to allocate). This method that simply sets all
stored contacts to be inactive, could do the same thing, but in a
fraction of the time.

Co-authored-by: Dorian Stoll <dorian.stoll@tmsp.io>