Add run count option to iptsd-perf
This allows changing the number of runs when working on detectors whose
running times are different orders of magnitude like the basic detector
compared to the advanced detector.

The idea was taken from hyperfine. There is no plan to add varying
numbers of runs. This option does not replace hyperfine because we need
a custom iptsd-perf to keep the input data and the detectors loaded in
memory while hyperfine launches new processes each time.