contacts: basic: Use local maximas for spanning clusters
Using local maximas avoids multiple iterations over the entire heatmap
and allows to only look closer at the spots that are interesting.

This change also moves the visited state to be per-cluster. This means,
that there can now be overlapping contacts.

A third change is that the cluster spanning search will stop, if the
value of the pixel is larger than the previous value. This prevents
creating clusters that span multiple connected contacts.

This change was inspired by and is based on the advanced blob detector
from @qzed and the work from @catphish who wrote their own touch
processor a while ago.