ipts: Only process the last element of a stylus report
The hardware seems to generate stylus data very fast, so
IPTS will cache all reports within a 5 millisecond window
and send them all at once.

If we send all that data at once we will mix data from
the past and the present, which leads to jitter in the
stylus movement.