contacts: validation: Remove distance checking.
This actually doesn't make a lot of sense, because once a contact is very close to a palm,
the cluster merging will already merge it into the palm contact.

And, generally speaking, the new algorithm is much more robust, so we should try to simplify
the other algorithms by removing obsolete safety guards.