contacts: stability: Remove "lasso" stabilization again
I don't completely understand what it does, but if I am not wrong all it does
is add a more or less permanent offset to the contact position.

So either I have no idea what this is, or it doesn't really help much with
stabilizing inputs. Either way, it's probably better to keep things simple.