contacts: Fix incorrect usages of casts::to
The contact library is filled with code that casts floating point numbers,
du to its templated nature. However, casts::to cannot be used to cast
doubles to floats because it cannot guarantee that input and output are
identical.

I did not notice this earlier because the contact library is never
instantiated with floats in iptsd, so the relevant assertion was not triggered.