core: linux: Allow Device sections in every file
We already allow this for files in the preset directory, so we can just
extend it to all config files that we load. This is going to be
necessary for the Surface Laptop Studio 2 and newer devices, where the
touchpad needs to be driven by iptsd and might need a different
calibration data than the normal touchscreen.